requirements.txt
cassandra-driver==3.12.0
matplotlib==2.1.2
pandas==0.22.0
pip==9.0.1
requests==2.18.4
sklearn==0.0
scipy==1.0.0
tensorflow==1.4.1
